+++ Menu.cpp
@@ -5,11 +5,11 @@
 class EasyLevelButton : public UIButton
 {
 public:
-	EasyLevelButton() : UIButton("Easy", { 412, 162}, { 200, 75 }) { }
+	EasyLevelButton() : UIButton("Easy", { 412.f, 162.f}, { 200.f, 75.f }) { }
 	void MouseButtonReleased(sf::Event e)
 	{
 		if (e.mouseButton.button == sf::Mouse::Button::Left //if the left mouse button was clicked
-			&& this->background.getGlobalBounds().contains(e.mouseButton.x, e.mouseButton.y)) //if the click was inside the button
+			&& this->background.getGlobalBounds().contains(static_cast<float>(e.mouseButton.x), static_cast<float>(e.mouseButton.y))) //if the click was inside the button
 		{
 			DifficultySettings::setDifficulty(DifficultySettings::EASY);
 			mainMenu.startTestLevel();
@@ -19,11 +19,11 @@ public:
 
 class NormalLevelButton : public UIButton{
 public:
-    NormalLevelButton() : UIButton("Normal", {412, 242}, {200, 75}) {}
+    NormalLevelButton() : UIButton("Normal", {412.f, 242.f}, {200.f, 75.f}) {}
     void MouseButtonReleased(sf::Event e)
     {
         if (e.mouseButton.button == sf::Mouse::Button::Left //if the left mouse button was clicked
-            && this->background.getGlobalBounds().contains(e.mouseButton.x, e.mouseButton.y)) //if the click was inside the button
+            && this->background.getGlobalBounds().contains(static_cast<float>(e.mouseButton.x), static_cast<float>(e.mouseButton.y))) //if the click was inside the button
         {
 			DifficultySettings::setDifficulty(DifficultySettings::NORMAL);
             mainMenu.startTestLevel();
@@ -33,11 +33,11 @@ public:
 
 class HardLevelButton : public UIButton{
 public:
-    HardLevelButton() : UIButton("Hard", {412, 322}, {200, 75}) {}
+    HardLevelButton() : UIButton("Hard", {412.f, 322.f}, {200.f, 75.f}) {}
     void MouseButtonReleased(sf::Event e)
     {
         if (e.mouseButton.button == sf::Mouse::Button::Left //if the left mouse button was clicked
-            && this->background.getGlobalBounds().contains(e.mouseButton.x, e.mouseButton.y)) //if the click was inside the button
+            && this->background.getGlobalBounds().contains(static_cast<float>(e.mouseButton.x), static_cast<float>(e.mouseButton.y))) //if the click was inside the button
         {
 			DifficultySettings::setDifficulty(DifficultySettings::HARD);
             mainMenu.startTestLevel();
@@ -55,11 +55,11 @@ public:
 class QuitButton : public UIButton
 {
 public:
-	QuitButton() : UIButton("Quit", { 412, 602}, { 200, 75 }) { }
+	QuitButton() : UIButton("Quit", { 412.f, 602.f}, { 200.f, 75.f }) { }
 	void MouseButtonReleased(sf::Event e)
 	{
 		if (e.mouseButton.button == sf::Mouse::Button::Left //if the left mouse button was clicked
-			&& this->background.getGlobalBounds().contains(e.mouseButton.x, e.mouseButton.y)) //if the click was inside the button
+			&& this->background.getGlobalBounds().contains(static_cast<float>(e.mouseButton.x), static_cast<float>(e.mouseButton.y))) //if the click was inside the button
 		{
 			this->screen->close();
 		}
