+++ TestLevel.h
@@ -32,7 +32,7 @@ public:
 		levelScreen.addMap(&map);
 
 		sf::Sprite m_Sprite;
-		m_Sprite.setPosition(Screen::windowWidth / 2, Screen::windowHeight / 2);
+		m_Sprite.setPosition(static_cast<float>(Screen::windowWidth / 2), static_cast<float>(Screen::windowHeight / 2));
 		static sf::Texture m_Texture;
 		m_Texture.loadFromFile("zombie.png");
 		m_Sprite.setTexture(m_Texture);
